Recurrence Rule Form v1
This commit represents the initial version of the Recurrence Rule form.
Previous iterations of this component suffered from some rendering
issues. This version does not have those issues, although it may be a
little longer than previous versions.

This version also allows users to specify an "initial" time object. This
may open up the possibility of editing event times in the future.