User Form v1
This commit adds v1 of the User Form, which will be used to allow a user
to change his/her name, phone #, and profile picture.

Profile images are validated using the JavaScript Image()
API. It loads the function on mount and checks to make sure that the
image URL provided loads. If so, the image is valid.

Since there's no way to pass this result directly to the error()
function, we'll need to assign a variable to "true" or "false" and check
this variable's state in the error function.

This validation is debounced, since it makes a network request every
time it sets the image source.