Map v1 + tests
This commit adds the initial version of the map. It shows markers for
meetups along with a sidebar with the number of meetups on the map. It
also allows you to zoom in on a country by specifying the country code
in the URL, e.g. "map/FRA" for France.

You can click on a marker to see a few details about the meetup in
question.