improve User Form
- run reset every time "initial" changes. It will only change when
submissions are complete, so for Profiles, etc. the user will see
results reflected immediately on the form when the parent component
changes the initial values.

- provide a "resetOnSubmit" parameter which resets the form on
submission. This is optional, and by default it doesn't. This prevents a
form from "resetting" to old values before database queries finish and
the "initial" values are changed

- move "disabled" to the bottom of the script to prevent strange
re-enablings of the submit button due to order of evaluation.