add User Location Form
This commit adds the user location form, allowing users to add their
location to the map. They can edit this location from their profile or
from the map itself.

User Locations are obscured by reducing the significant digits on the
geolocation.

This commit also updates the User Card to allow for simplified location
showing.

It creates a different-colored Map marker for users vs. meetups.

It also updates the Cypress tests.

Since user profiles can now be accessed from the map, I needed to clean
up the fields made available for anonymous users:

- change the profileData helper to accept an 'authorize' parameter. This
prevents unauthorized users from requesting a user's games and phone #.
- change the non-profile user page to make games optional. Also include
a prompt that encourages the user to log in to see games.
- add an additional "isCurrentUser" parameter to the profileData helper
method. This is to prevent other users from being able to look at the
browser requests and determine a user's address.