create refresh link for events
This commits adds the 'refresh' path for events, enabling users to
refresh the 'last updated' time on their events in one click. Ideally
this link would be sent in an email to users with meetups which haven't
been updated in six months. That's coming later...

Also move the event page and attendance helper into its own [id]
directory to facilitate creation of the refresh route.