add location data API file
This commit adds the location data API file, /api/location/index. It
also moves the country data from the Map routes directory to this new
API directory, which makes more sense. The Location Data API contains
two methods:

- getBoundingBox, which gets the bounding box given a country code. The
/map/[country] route calls this function to retrieve the bounding box.
- getCountry, which uses ipinfo.io to find the user's country for
automatic centering of the map. This will be used in future versions of
the map to automatically center on the user's country when they navigate
to the map without a country in mind.