add map auto-centering
This commit adds a new Map control: centerOnUser. If centerOnUser is
true and no bounds are provided to Mapbox, the map will use the IPInfo
service to determine the user's country and center the map on it. You
can always turn that off with centerOnUser = false, and it will still
center on the US.

This commit also moves a couple of utility functions, such as checking
the bounds and adding articles to the name of a country, to a utility
function file, /utils/location.js